Extracted few things to a shared folder
