+++ src/Parser/Warlock/Affliction/CombatLogParser.js
@@ -22,13 +22,13 @@ import MaleficGrasp from './Modules/Talents/MaleficGrasp';
 import Contagion from './Modules/Talents/Contagion';
 import AbsoluteCorruption from './Modules/Talents/AbsoluteCorruption';
 import EmpoweredLifeTap from './Modules/Talents/EmpoweredLifeTap';
-import SoulHarvest from './Modules/Talents/SoulHarvest';
-import SoulHarvestTalent from './Modules/Talents/SoulHarvestTalent';
+import SoulHarvest from '../Shared/Modules/Talents/SoulHarvest';
+import SoulHarvestTalent from '../Shared/Modules/Talents/SoulHarvestTalent';
 import DeathsEmbrace from './Modules/Talents/DeathsEmbrace';
 import SiphonLifeUptime from './Modules/Talents/SiphonLifeUptime';
 import SoulConduit from './Modules/Talents/SoulConduit';
 
-import TheMasterHarvester from './Modules/Items/Legendaries/TheMasterHarvester';
+import TheMasterHarvester from '../Shared/Modules/Items/TheMasterHarvester';
 import StretensSleeplessShackles from './Modules/Items/Legendaries/StretensSleeplessShackles';
 import SoulOfTheNetherlord from './Modules/Items/Legendaries/SoulOfTheNetherlord';
 import PowerCordOfLethtendris from './Modules/Items/Legendaries/PowerCordOfLethtendris';
