+++ src/Parser/Warlock/Demonology/Modules/SoulShards/SoulShardDetails.js
@@ -4,10 +4,10 @@ import Analyzer from 'Parser/Core/Analyzer';
 import Tab from 'Main/Tab';
 import StatisticBox, { STATISTIC_ORDER } from 'Main/StatisticBox';
 
+import WastedShardsIcon from 'Parser/Warlock/Shared/Images/warlock_soulshard_bw.jpg';
+
 import SoulShardBreakdown from './SoulShardBreakdown';
 import SoulShardTracker from './SoulShardTracker';
-
-import WastedShardsIcon from '../../Images/warlock_soulshard_bw.jpg';
 
 const soulShardIcon = 'inv_misc_gem_amethyst_02';
