+++ src/Parser/Warlock/Destruction/CombatLogParser.js
@@ -22,7 +22,7 @@ import ReverseEntropy from './Modules/Talents/ReverseEntropy';
 import Eradication from './Modules/Talents/Eradication';
 import EmpoweredLifeTap from './Modules/Talents/EmpoweredLifeTap';
 import FireAndBrimstone from './Modules/Talents/FireAndBrimstone';
-import SoulHarvest from './Modules/Talents/SoulHarvest';
+import SoulHarvest from '../Shared/Modules/Talents/SoulHarvest';
 import SoulHarvestTalent from './Modules/Talents/SoulHarvestTalent';
 import ChannelDemonfire from './Modules/Talents/ChannelDemonfire';
 import TalentHub from './Modules/Talents/TalentHub';
@@ -30,11 +30,11 @@ import TalentHub from './Modules/Talents/TalentHub';
 import AlythesssPyrogenics from './Modules/Items/Legendaries/AlythesssPyrogenics';
 import FeretoryOfSouls from './Modules/Items/Legendaries/FeretoryOfSouls';
 import LessonsOfSpaceTime from './Modules/Items/Legendaries/LessonsOfSpaceTime';
-import SindoreiSpite from './Modules/Items/Legendaries/SindoreiSpite';
+import SindoreiSpite from '../Shared/Modules/Items/SindoreiSpite';
 import MagistrikeRestraints from './Modules/Items/Legendaries/MagistrikeRestraints';
 import OdrShawlOfTheYmirjar from './Modules/Items/Legendaries/OdrShawlOfTheYmirjar';
 import SoulOfTheNetherlord from './Modules/Items/Legendaries/SoulOfTheNetherlord';
-import TheMasterHarvester from './Modules/Items/Legendaries/TheMasterHarvester';
+import TheMasterHarvester from '../Shared/Modules/Items/TheMasterHarvester';
 
 import T20_2set from './Modules/Items/T20_2set';
 import T21_2set from './Modules/Items/T21_2set';
