+++ src/Parser/Warlock/Destruction/Modules/Talents/TalentHub.js
@@ -6,7 +6,6 @@ import Analyzer from 'Parser/Core/Analyzer';
 
 //credit to hpal mod, i modified the traits script
 import Backdraft from './Backdraft';
-import SoulHarvestTalent from './SoulHarvestTalent';
 import ChannelDemonfire from './ChannelDemonfire';
 import Eradication from './Eradication';
 import EmpoweredLifeTap from './EmpoweredLifeTap';
@@ -14,34 +13,33 @@ import FireAndBrimstone from './FireAndBrimstone';
 import ReverseEntropy from './ReverseEntropy';
 import RoaringBlaze from './RoaringBlaze';
 import Shadowburn from './Shadowburn';
+import SoulHarvestTalent from './SoulHarvestTalent';
 
 class TalentHub extends Analyzer {
   static dependencies = {
     backdraft: Backdraft,
-    soulHarvestTalent: SoulHarvestTalent,
+    shadowburn: Shadowburn,
+    roaringBlaze: RoaringBlaze,
     eradication: Eradication,
     empoweredLifeTap: EmpoweredLifeTap,
+    reverseEntropy: ReverseEntropy,
+    fireAndBrimstone: FireAndBrimstone,
+    soulHarvestTalent: SoulHarvestTalent,
     channelDemonfire: ChannelDemonfire,
-    fnb: FireAndBrimstone,
-    entropy: ReverseEntropy,
-    roaringBlaze: RoaringBlaze,
-    shadowburn: Shadowburn,
   };
 
   statistic() {
     return (
-      <StatisticsListBox
-        title="Talents"
-      >
+      <StatisticsListBox title="Talents">
         {this.backdraft.active && this.backdraft.subStatistic()}
+        {this.shadowburn.active && this.shadowburn.subStatistic()}
+        {this.roaringBlaze.active && this.roaringBlaze.subStatistic()}
         {this.eradication.active && this.eradication.subStatistic()}
         {this.empoweredLifeTap.active && this.empoweredLifeTap.subStatistic()}
-        {this.channelDemonfire.active && this.channelDemonfire.subStatistic()}        
+        {this.reverseEntropy.active && this.reverseEntropy.subStatistic()}
+        {this.fireAndBrimstone.active && this.fireAndBrimstone.subStatistic()}
         {this.soulHarvestTalent.active && this.soulHarvestTalent.subStatistic()}
-        {this.fnb.active && this.fnb.subStatistic()}
-        {this.entropy.active && this.entropy.subStatistic()}
-        {this.roaringBlaze.active && this.roaringBlaze.subStatistic()}
-        {this.shadowburn.active && this.shadowburn.subStatistic()}
+        {this.channelDemonfire.active && this.channelDemonfire.subStatistic()}
       </StatisticsListBox>
     );
   }
