+++ src/Parser/Warlock/Shared/Modules/Items/SindoreiSpite.js
