+++ src/Parser/Warlock/Shared/Modules/Talents/SoulHarvest.js
