+++ src/Parser/Warlock/Shared/Modules/Talents/SoulHarvestTalent.js
@@ -2,10 +2,11 @@ import React from 'react';
 
 import Analyzer from 'Parser/Core/Analyzer';
 import Combatants from 'Parser/Core/Modules/Combatants';
+
 import SPELLS from 'common/SPELLS';
 import SpellIcon from 'common/SpellIcon';
 import { formatNumber, formatPercentage } from 'common/format';
-import StatisticBox, { STATISTIC_ORDER } from 'Main/StatisticBox';
+import StatisticBox from 'Main/StatisticBox';
 
 import SoulHarvest from './SoulHarvest';
 
@@ -30,8 +31,6 @@ class SoulHarvestTalent extends Analyzer {
       />
     );
   }
-
-  statisticOrder = STATISTIC_ORDER.OPTIONAL(2);
 }
 
 export default SoulHarvestTalent;
